Add a small trick to improve mismatch kernel computation
Example

| i |id_seq| seq | count |
|---|------|-----|-------|
| 0 |    0 | ATA |     2 |
| 1 |    0 | AAA |     4 |
| 2 |    0 | ATT |     3 |
| 3 |    1 | ATT |     5 |
| 4 |    1 | TTA |     1 |

if we iterate from i=0..N and j=i..N (N = 5 in example), instead of
i=0..N and j=0..N, we compute the upper triangle of the matrix but we
need to be careful for diagonal elements K(i, i), because of the
change in the second loop, for such coefficients, we need to multiply
the number of matches by 2. This works well because the array
is sorted by sequence id.